iotparkingsystem/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
gunicorn==22.0.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
packaging==24.1
pyserial==3.5
sqlparse==0.5.0
tzdata==2024.1
